LRP-4764 We don't need to update release tool to build portal

+++ release/_git.sh
@@ -131,6 +131,13 @@ function update_portal_repository {
 function update_release_tool_repository {
 	trap 'return ${LIFERAY_COMMON_EXIT_CODE_BAD}' ERR
 
+	if [ "${LIFERAY_RELEASE_PRODUCT_NAME}" == "portal" ]
+	then
+		lc_log INFO "The product is set to portal."
+
+		return "${LIFERAY_COMMON_EXIT_CODE_SKIPPED}"
+	fi
+
 	lc_cd "${_PROJECTS_DIR}"/liferay-release-tool-ee
 
 	git reset --hard && git clean -dfx
